tools: write tailscale to hosts instead of ssh_config

+++ install.sh
@@ -8,13 +8,15 @@ function append() {
   local text="$1" file="$2"
 
   if ! grep -q "$text" "$file"; then
-    echo -e "$text" >> "$file"
+    echo "$text" >>"$file"
   fi
 }
 
+echo "fetching public key..."
 mkdir -p $HOME/.ssh
 append "$(curl https://github.com/tobi.keys)" $HOME/.ssh/authorized_keys
 
+echo "installing dotfiles..."
 if test -d $HOME/dotfiles; then
   echo "dotfiles exists"
 else
@@ -23,19 +25,25 @@ fi
 
 cd $HOME/dotfiles
 
-# install shell hooks
+echo "installing dotfiles to .zshrc..."
 hook="source ~/dotfiles/shell"
 append "$hook" $HOME/.zshrc
 
-if command -v 'zsh' &> /dev/null; then
-  cd $HOME && exec zsh
-else
+if command -v 'zsh' &>/dev/null; then
   echo "* install zsh!"
-  sudo apt install zsh
-fi
 
-echo "chsh -s $(which zsh)"
-chsh -s "$(which zsh)"
+  echo "Do you want to change your shell to zsh? (y/n)"
+  read answer
+  if echo "$answer" | grep -iq "^y"; then
+    echo "Changing shell to zsh..."
+    sudo apt install zsh
+    chsh -s "$(which zsh)"
+    echo "Shell changed to zsh."
+  else
+    echo "Shell not changed."
+  fi
+
+fi
 
 echo
 echo "done"
